+++ mouse_base/src/MouseMovement.h
@@ -10,7 +10,7 @@ struct Overshoot {
     int right;
 };
 
-constexpr unsigned MOVEMENT_SPEED = 100;
+constexpr unsigned short MOVEMENT_SPEED = 100;
 
 class MouseMovement {
     bool movementInProgress = false;
@@ -24,12 +24,24 @@ class MouseMovement {
     bool prevLeftReading = false;
     bool prevRightReading = false;
 public:
+    /// Sets the left/right motors to run at the given values
     static void runMotors(int left, int right);
 
+    /// Runs the motors until they have moved the given number of steps
+    /// <returns>Returns the overshoot of the target steps</returns>
     static Overshoot moveStepsBlocking(int stepsLeft, int stepsRight);
 
+    /// Starts a movement that will run the motors until they have moved the given number of steps
+    ///
+    /// Requires updateMovement to be called in a tight loop until the movement is complete
     void moveStepsNonBlocking(int stepsLeft, int stepsRight);
 
+    /// Moves the mouse until the current movement is completed
+    ///
+    /// !WARNING! If not called frequently enough the step measurement will be innaccurate! Ensure that no blocking code
+    /// is run in the loop that contains this method call
+    ///
+    /// <returns>false when the movement is completed</returns>
     bool updateMovement();
 };
